pylibbe: cast between incompatible function types
../common/libbe_py.c:1091:13: error: cast between incompatible function types fr
om 'PyObject * (*)(PyObject *, PyObject *, PyObject *)' {aka 'struct _object * (
*)(struct _object *, struct _object *, struct _object *)'} to 'PyObject * (*)(Py
Object *, PyObject *)' {aka 'struct _object * (*)(struct _object *, struct _obje
ct *)'} [-Werror=cast-function-type]
  {"beList", (PyCFunction)beList, METH_VARARGS | METH_KEYWORDS,
             ^
cc1: all warnings being treated as errors

+++ usr/src/lib/pylibbe/common/libbe_py.c
@@ -1088,7 +1088,7 @@ static struct PyMethodDef libbeMethods[] = {
 	    "Destroy a snapshot."},
 	{"beMount", (PyCFunction)beMount, METH_VARARGS, "Mount a BE."},
 	{"beUnmount", (PyCFunction)beUnmount, METH_VARARGS, "Unmount a BE."},
-	{"beList", (PyCFunction)beList, METH_VARARGS | METH_KEYWORDS,
+	{"beList", (PyCFunction)(uintptr_t)beList, METH_VARARGS | METH_KEYWORDS,
 	    "List BE info."},
 	{"beRename", (PyCFunction)beRename, METH_VARARGS, "Rename a BE."},
 	{"beActivate", (PyCFunction)beActivate, METH_VARARGS, "Activate a BE."},
